refactor(ImportExport): Moved crypt function to WalletCrypto
